refactor(sync-actions): to rename create action map custom
