+++ packages/sync-actions/src/cart-discounts.js
@@ -1,7 +1,7 @@
 import flatten from 'lodash.flatten'
 import createBuildActions from './utils/create-build-actions'
 import createMapActionGroup from './utils/create-map-action-group'
-import actionsMapCustom from './utils/create-map-action-custom'
+import actionsMapCustom from './utils/action-map-custom'
 import { actionsMapBase } from './cart-discounts-actions'
 import * as diffpatcher from './utils/diffpatcher'
