+++ packages/sync-actions/src/categories.js
@@ -3,7 +3,7 @@ import type { SyncAction, ActionGroup } from 'types/sdk'
 import flatten from 'lodash.flatten'
 import createBuildActions from './utils/create-build-actions'
 import createMapActionGroup from './utils/create-map-action-group'
-import actionsMapCustom from './utils/create-map-action-custom'
+import actionsMapCustom from './utils/action-map-custom'
 import * as categoryActions from './category-actions'
 import * as diffpatcher from './utils/diffpatcher'
